+++ docs/4.5.1-release-notes.md
@@ -1,2 +1,5 @@
 ### Fix
 - Prevent the REQUEST_DENIED response caused by ZERO_RESULTS in the Place ID API call
+
+### Chore
+- Code clean up
