+++ src/ExternalSearch.Providers.GoogleMaps/Models/Location/LocationDetailsResponse.cs
@@ -1,0 +1,98 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.GoogleMaps.Models.Location;
+public class AddressComponent
+{
+
+    [JsonProperty("long_name")]
+    public string LongName { get; set; }
+
+    [JsonProperty("short_name")]
+    public string ShortName { get; set; }
+
+    [JsonProperty("types")]
+    public List<string> Types { get; set; }
+}
+
+public class Location
+{
+
+    [JsonProperty("lat")]
+    public string Lat { get; set; }
+
+    [JsonProperty("lng")]
+    public string Lng { get; set; }
+}
+
+public class Northeast
+{
+
+    [JsonProperty("lat")]
+    public double Lat { get; set; }
+
+    [JsonProperty("lng")]
+    public double Lng { get; set; }
+}
+
+public class Southwest
+{
+
+    [JsonProperty("lat")]
+    public string Lat { get; set; }
+
+    [JsonProperty("lng")]
+    public string Lng { get; set; }
+}
+
+public class Viewport
+{
+
+    [JsonProperty("northeast")]
+    public Northeast Northeast { get; set; }
+
+    [JsonProperty("southwest")]
+    public Southwest Southwest { get; set; }
+}
+
+public class Geometry
+{
+
+    [JsonProperty("location")]
+    public Location Location { get; set; }
+
+    [JsonProperty("viewport")]
+    public Viewport Viewport { get; set; }
+}
+
+public class Result
+{
+
+[JsonProperty("address_components")]
+public List<AddressComponent> AddressComponents { get; set; }
+
+[JsonProperty("adr_address")]
+public string AdrAddress { get; set; }
+
+[JsonProperty("formatted_address")]
+public string FormattedAddress { get; set; }
+
+[JsonProperty("geometry")]
+public Geometry Geometry { get; set; }
+
+[JsonProperty("name")]
+public string Name { get; set; }
+}
+
+public class LocationDetailsResponse
+{
+
+    [JsonProperty("html_attributions")]
+    public List<object> HtmlAttributions { get; set; }
+
+    [JsonProperty("result")]
+    public Result Result { get; set; }
+
+    [JsonProperty("status")]
+    public string Status { get; set; }
+}
