+++ src/ExternalSearch.Providers.GoogleMaps/Vocabularies/GoogleMapsVocabulary.cs
@@ -1,20 +1,16 @@
-﻿using CluedIn.Core.Data;
-using CluedIn.Core.Data.Vocabularies;
+﻿namespace CluedIn.ExternalSearch.Providers.GoogleMaps.Vocabularies;
 
-namespace CluedIn.ExternalSearch.Providers.GoogleMaps.Vocabularies
+public static class GoogleMapsVocabulary
 {
-    public static class GoogleMapsVocabulary
+    /// <summary>
+    /// Initializes static members of the <see cref="GoogleMapsVocabulary" /> class.
+    /// </summary>
+    static GoogleMapsVocabulary()
     {
-        /// <summary>
-        /// Initializes static members of the <see cref="KnowledgeGraphVocabulary" /> class.
-        /// </summary>
-        static GoogleMapsVocabulary()
-        {
-            Organization = new CompanyDetailsVocabulary();
-            Location = new LocationDetailsVocabulary();
-        }
-
-        public static CompanyDetailsVocabulary Organization { get; private set; }
-        public static LocationDetailsVocabulary Location { get; private set; }
+        Organization = new CompanyDetailsVocabulary();
+        Location = new LocationDetailsVocabulary();
     }
+
+    public static CompanyDetailsVocabulary Organization { get; private set; }
+    public static LocationDetailsVocabulary Location { get; private set; }
 }
